@@ -2,7 +2,9 @@ package com.example.obrestdatajpa.entities;
 
 
 
-import javax.persistence.*;
+
+import jakarta.persistence.*;
+
 import java.time.LocalDate;
 
 @Entity
